Repositories ID changed from Integer to Long

+++ src/main/java/com/uep/wap/repository/ChatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChatRepository extends CrudRepository<Chat, Integer> {
+public interface ChatRepository extends CrudRepository<Chat, Long> {
 }
